Get new-path-transform-of working
This is the algorithm using heaps to compute the
path transform in time nearly linear in the number of
nodes unique to either tree.

For now, this is only invoked in an around method for the
old algorithm.  The around method compares the results from
both and asserts their equivalence.