Add new path-transform-of algorithm
This is executed in parallel with the old algorithm,
for testing, but will be switched over soon.

Unconditionalize some tests that are now passing.
There is an issue with the implementation of SWAP that
needs thinking about.  The issue is that a rewrite
that occurs in stages can lose information.  The
probably solution is to arrange the intermediate rewrites
to be forgotten when computing transforms.