Add rewrite encapsulation
This allows a transformation on a tree to group up all its changes
into a single change, as reflected by the TRANSFORMS slot of the root.
This is a good idea if the intermediate trees are not internally
consistent, or if a node is removed and then reinserted.  An example
of this is the SWAP operation, where encapsulation is naturally
performed in an around method.