Switching traverse-nodes to do-tree, test rework
This commit *significantly* re-works our test suite by changing the
type of node-with-data to node-cons, and updating our tests so that
they faithfully reflect the new intended usage pattern (child-slots).
This resulted in *many* changes through-out the test suite.  In many
cases tests are still broken and are marked as expected to fail.
These need to be re-worked.  In some cases the actual implementation
needs work and in other cases the test just needs to be adapted.