Translate README.md to README.org
By using Org in CI, we'll be able to keep this README in sync with the
code as both evolve over time. As Eric pointed out in our Scrum
discussion earlier today:

> I'm tempted to say we could translate the README to org-mode and then
> use org-mode's code block support to do this for us, as that was my
> baby in a former life.
> https://eschulte.github.io/data/CISE-13-3-SciProg.pdf
> https://orgmode.org/manual/Working-with-Source-Code.html

I based this translation on [a GitLab Pages example][1]. To check that I
was using the right header levels (I wasn't sure whether or not to
subtract one from the number of octothorpes to get the number of
asterisks), I downloaded the HTML source from that page and from the
existing functional-trees README page on GitLab in order to see that,
for instance, `##` and `**` both correspond to `<h2>`.

As a sanity check about the language name after `#+BEGIN_SRC`, I ran
`(use-package htmlize)` and used [`C-c C-e h h`][2] to view the output
in my browser. This again maps `**` to `<h2>`, but it also creates a
table of contents, numbers each header, and lists some author and
creation date information at the bottom. I'm going by the aforementioned
GitLab example to guess that these won't show up when I push this.

[1]: https://gitlab.com/pages/org-mode/-/blob/8ef8d6b21fe1068875f2d54093ee9f06809639d7/README.org
[2]: https://orgmode.org/manual/HTML-export-commands.html