Add Python script to run Lisp examples from README
I plan to soon add this script to the CI setup.

GitLab [actually uses][1] the Ruby library [CommonMarker][2] to process
Markdown files in repositories. Because CommonMark is standardized, I
chose to instead use the Python CommonMark parser [Marko][3], because
Python is one of the three languages we're using for this larger
project. I chose Marko over the [other Python CommonMark parsers][4]
because it seemed to most directly support GFM, which should correspond
to the same set of extensions that GitLab uses via CommonMarker.

I initially tried using the built-in Python [`subprocess`][5] module,
but that [doesn't work well for multiple interactions with the REPL][6],
so I chose to use the Pexpect library instead. Pexpect does have a
specific [`replwrap`][7] module that looks like it should be perfect for
this use case, but I wasn't able to get that to work immediately, so I
just reverted to my working solution that uses `pexpect.spawn`.

[1]: https://docs.gitlab.com/ee/user/markdown.html#transition-from-redcarpet-to-commonmark
[2]: https://github.com/gjtorikian/commonmarker
[3]: https://github.com/frostming/marko
[4]: https://github.com/commonmark/commonmark-spec/wiki/List-of-CommonMark-Implementations#python
[5]: https://stackoverflow.com/a/14458956
[6]: https://stackoverflow.com/a/28616322
[7]: https://pexpect.readthedocs.io/en/stable/api/replwrap.html