Give better error messages from readme.py
Previously, if the output didn't match, the actual and expected output
strings would be printed in a Python-readable format, which made it very
difficult to see at a glance what differed between them since newlines
were escaped. Also, unnecessarily long error messages were printed on
EOF and timeout, so I replaced them with shorter ones that give the same
relevant information.

As a side note, the previous file header docstring didn't mention the
EOF case, so I added that in as well.