Install directly rather than from requirements.txt
The issue with my previous attempts was that they assumed the Dockerfile
was automatically being used to build a new Docker image, which it
wasn't. I needed to [build and push a new Docker image][1]:

    docker login docker.grammatech.com
    docker build -t docker.grammatech.com/synthesis/functional-trees .
    docker push docker.grammatech.com/synthesis/functional-trees

Then once I started following those steps, I realized that of course the
Docker image gets built before the repository contents are available, so
the Dockerfile can't make use of `requirements.txt`, and must instead
install all the Python dependencies explicitly, hence this commit.

[1]: https://git.grammatech.com/help/user/packages/container_registry/index