Lengthen readme.py timeout from 1 second to 30
The one-second timeout is nice for local development when all the Lisp
libraries are already downloaded and built, but in CI, it takes more
than one second to evaluate `(ql:quickload :functional-trees)`. Thus for
simplicity's sake, I'm removing the argument so it just uses the
[Pexpect default][1] of 30 seconds.

This shouldn't matter too much for local development anyway, because it
will still be eventually correct (even more likely than it was before),
and a developer running `./readme.py` at the command line can just hit
`^C` if they see that it triggered the debugger.

[1]: https://pexpect.readthedocs.io/en/stable/api/pexpect.html#spawn-class