Discuss fingers
This behavior of `ft:populate-fingers` seems wrong; I don't see a
function in functional-trees which takes a `path` of this form and
actually performs the lookup to compute the node at the end of the path.
From what I've heard in our more recent design discussions, it sounds
like nowadays a path is just supposed to be a list of nonnegative
integers corresponding to indices in the `children` of a node.