Make path transform computation work generally
The problem is that a subtree with fingers already present
can be pasted inside another tree.  Transforms from those
nodes are not computable using the predecessor chain of the
root of the tree, but must be computed from scratch.  Do that
as a fallback.   This can cause efficiency problems in general;
if that becomes a problem we'll need to address it by caching.