Various tweaks
Make searches splay the tree, to avoid O(n^2) cases.
Reorder arguments to find functions for consistency.
Enable :serial-number argument of NIL to allocate
a new serial number.  This is useful to avoid SN collisions
when a rewrite sinks a term inside a newly constructed term,
as happens in wrap edits in resolve.